+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 Changelog
 =========
 
+### 0.14.0
+ - method calls (`#...` in variable expressions) supported, though they don't
+   really do anything as their are no expression objects in the Thymeleaf
+   context just yet
+
 ### 0.13.1
  - template names in fragment expressions can have `/` in their name (as they
    are often paths to a file)